pyproject.toml
[tool.poetry]
name = "iperocks-croqui-ui"
version = "0.1.0"
description = ""
authors = ["Jean Martins <jean.martins@ericsson.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.37.0"
pdf2image = "^1.17.0"
pdfplumber = "^0.11.2"
pillow = "^10.4.0"
pytesseract = "^0.3.10"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
unidecode = "^1.3.8"
streamlit-antd-components = "^0.3.2"
fpdf = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
